build-system/common/diff.js
